Split off xn297 code.
