+++ Makefile
@@ -427,6 +427,11 @@ COMMON_SRC = \
             io/statusindicator.c \
             rx/ibus.c \
             rx/msp.c \
+            rx/nrf24.c \
+            rx/nrf24_cx10.c \
+            rx/nrf24_syma.c \
+            rx/nrf24_v202.c \
+            rx/nrf24_h8_3d.c \
             rx/pwm.c \
             rx/rx.c \
             rx/sbus.c \
